Remove Chroma styling choices.
Remove some default styling choices that Chroma seems to have inherited
from Pygments. The Pygments team have made inconsistent choices from
theme to theme that don't make sense for Base16. The classes map is
easily seen in the file, so we'll leave it to the user on all but the
few items that have an obvious consistent choice.